+++ Lab4/df_manip.py
@@ -1,9 +1,6 @@
 import argparse
 import pandas as pd
 import csv
-from nltk.stem import WordNetLemmatizer
-from nltk.tokenize import word_tokenize
-import matplotlib.pyplot as plt
 import logging
 import os
 
@@ -50,20 +47,6 @@ def group(data: pd.DataFrame) -> pd.DataFrame:
     return data.groupby('stars').agg({'count words': ['max', 'min', 'mean']})
 
 
-def make_hist(df: pd.DataFrame, star: int) -> list:
-    """
-    does lemmatic analysis
-    """
-    lemmatizer = WordNetLemmatizer()
-    lemmas = []
-    data = filter_by_class(df, star)
-    for i in range(len(data)):
-        for word in word_tokenize(data['review'].values[i]):
-            if ord(word[0]) > 191:
-                lemmas.append(lemmatizer.lemmatize(word))
-    return lemmas
-
-
 def get_statistic(data: pd.DataFrame) -> pd.DataFrame:
     """
     get statistic about numeric cols
@@ -74,18 +57,9 @@ def get_statistic(data: pd.DataFrame) -> pd.DataFrame:
         logging.error('error in get_statistic')
 
 
-def draw_hists(df: pd.DataFrame) -> None:
-    """
-    draw histogramm
-    """
-    plt.bar(range(1, 6), [len(make_hist(df, i)) for i in range(1, 6)])
-    plt.legend()
-    plt.show()
-
-
 if __name__ == '__main__':
     logging.basicConfig(level=logging.INFO, filename=os.path.join('py_log.log'), filemode='w')
     parser = argparse.ArgumentParser(description="Input csv path, label of class")
     parser.add_argument("-c", "--csv", help="Input csv path", type=str)
     args = parser.parse_args()
-    draw_hists(make_data_frame(args.csv))
+    make_data_frame(args.csv)
